feat: create an extended sprite base
